implement Inference
